+++ app/api/fileHandlers.ts
@@ -1,4 +1,4 @@
-import { Track } from '../app/api/db'
+import { Track } from '~/api/db'
 
 const _getFile = async (track: Track): Promise<File | undefined> => {
   let handle = track.dirHandle || track.fileHandle
