+++ app/components/mixes/trackform.tsx
@@ -2,14 +2,14 @@ import { useEffect, useRef, useState } from 'react'
 import { ButtonGroup, Dialog } from '@mui/material'
 import { AccessTime, Eject, Pause, PlayArrow, Undo } from '@mui/icons-material'
 import { Button, Card, TextField } from '@mui/joy'
-import Loader from '../../layout/loader'
+import Loader from '~/styles/loader'
 import Slider, { SliderProps } from 'rc-slider'
-import { initPeaks } from '../../../app/api/initPeaks'
+import { initPeaks } from '~/api/initPeaks'
 import { PeaksInstance } from 'peaks.js'
-import Tracks from '../tracks/Tracks'
+import Tracks from '~/components/tracks/Tracks'
 import WaveformData from 'waveform-data'
-import { Track, db, TrackState, useLiveQuery } from '../../../app/api/db'
-import { Events } from '../../../app/api/Events'
+import { Track, db, TrackState, useLiveQuery } from '~/api/db'
+import { Events } from '~/api/Events'
 
 const TrackForm = ({ trackKey }: { trackKey: number }) => {
   interface SliderControlProps extends SliderProps {
