+++ package.json
@@ -28,6 +28,7 @@
     "cross-env": "^7.0.3",
     "dexie": "^3.2.2",
     "dexie-react-hooks": "^1.1.1",
+    "fake-indexeddb": "^4.0.0",
     "moment": "^2.29.4",
     "notistack": "^2.0.5",
     "peaks.js": "^2.0.6",
@@ -37,7 +38,7 @@
     "react-dom": "17.0.2",
     "remix-utils": "^3.3.0",
     "waveform-data": "^4.3.0",
-    "web-audio-beat-detector": "^8.1.33"
+    "web-audio-beat-detector": "^8.1.35"
   },
   "devDependencies": {
     "@remix-run/dev": "^1.7.2",
